+++ lib/view/page/book_detail/book_detail_page/book_detail_page.dart
@@ -10,6 +10,8 @@ import 'package:readme_app/core/constants/yh_style_icons.dart';
 import 'package:readme_app/model/book_detail_mock_data.dart';
 import 'package:readme_app/model/review_mock_data.dart';
 import 'package:readme_app/util/star_score/star_score.dart';
+import 'package:readme_app/view/page/book_detail/book_detail_page/components/membership_button.dart';
+import 'package:readme_app/view/page/book_detail/book_detail_page/components/purchase_button.dart';
 
 class BookDetailPage extends StatefulWidget {
   const BookDetailPage({Key? key}) : super(key: key);
@@ -775,54 +777,5 @@ class _BookDetailPageState extends State<BookDetailPage>
   }
 }
 
-class PurchaseButton extends StatelessWidget {
-  final String text;
-  final VoidCallback onPressed;
 
-  const PurchaseButton({
-    Key? key,
-    required this.text,
-    required this.onPressed,
-  }) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      style: ElevatedButton.styleFrom(
-        minimumSize: const Size(double.infinity, 50),
-        padding: const EdgeInsets.symmetric(horizontal: 10),
-        backgroundColor: Colours.app_sub_black,
-        foregroundColor: Colours.app_sub_white,
-        textStyle: const TextStyle(fontSize: 18),
-      ),
-      child: Text(text),
-      onPressed: onPressed,
-    );
-  }
-}
-
-class MembershipButton extends StatelessWidget {
-  final String text;
-  final VoidCallback onPressed;
-
-  const MembershipButton({
-    Key? key,
-    required this.text,
-    required this.onPressed,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      style: ElevatedButton.styleFrom(
-        minimumSize: const Size(double.infinity, 50),
-        padding: const EdgeInsets.symmetric(horizontal: 10),
-        backgroundColor: Colours.app_sub_black,
-        foregroundColor: Colours.app_sub_white,
-        textStyle: const TextStyle(fontSize: 18),
-      ),
-      child: Text(text),
-      onPressed: onPressed,
-    );
-  }
-}
