+++ lib/view/page/book_detail/book_detail_page/components/purchase_button.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:readme_app/core/constants/colours.dart';
+
+class PurchaseButton extends ConsumerWidget {
+  final String text;
+  final VoidCallback onPressed;
+
+  const PurchaseButton({
+    Key? key,
+    required this.text,
+    required this.onPressed,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return ElevatedButton(
+      style: ElevatedButton.styleFrom(
+        minimumSize: const Size(double.infinity, 50),
+        padding: const EdgeInsets.symmetric(horizontal: 10),
+        backgroundColor: Colours.app_sub_black,
+        foregroundColor: Colours.app_sub_white,
+        textStyle: const TextStyle(fontSize: 18),
+      ),
+      child: Text(text),
+      onPressed: onPressed,
+    );
+  }
+}
